+++ crates/sonar/src/prelude.rs
@@ -1,2 +1,1 @@
 pub use crate::error::*;
-pub use crate::specificity::{check, check_with_writer, mapping_to_paf};